refactor: move rendition group logic to HLS module
